Update patch set 2
Patch Set 2:

merged before I got chance to add a comment this afternoon.. 
While experimenting with the lenovo camera app that is doing the rounds;
 I noticed the Z works on 40 and 60 iso, 
50 doesn't show up, and 50 doesn't show up in Focal either using this patch. 
So it maybe worth altering 50 here to 40 and 60 ISO values instead

I haven't tested this in the arrays yet, but will do later. 
Thoughts?

Patch-set: 2