Update patch set 2
Patch Set 2:

Well, looking at this strictly as a user, the camera can now be used the first time you open it after booting, without having to worry about the aforementioned issue, and they wouldn't even realise that this is being done.


This could use a rename, if we reach a consensus on this.

Patch-set: 2