Update patch set 1
Patch Set 1:

Doesn't build (for hayabusa, using blue-common, msm8960-common and this repo)
because BoardConfigCommon.mk in blue-common uses ":=" to define TARGET_SPECIFIC_HEADER_PATH, and the variable set here is overriden by it. Maybe we should change := to += in all the series-common repos, but that's big amount of work to do and test. Doesn't know if there's some better way...

Patch-set: 1